fix: rename alpha to a
